requirements.txt
langchain>=0.1.9
langchain-community>=0.0.24
langchain-core>=0.1.27
langchain-openai>=0.0.7
python-dotenv>=1.0.1
streamlit>=1.31.0
pypdf2==3.0.1
docx2txt==0.8
pandas==2.1.3
pillow==10.1.0
pytesseract==0.3.10
chromadb==0.4.18
sentence-transformers==2.2.2
openai==1.3.0
azure-openai==1.1.0
python-magic-bin==0.4.14
tiktoken==0.5.1
faiss-cpu>=1.7.4
unstructured>=0.12.3
pdf2image>=1.17.0
python-docx>=1.1.0
python-magic>=0.4.27
openpyxl>=3.1.2
Pillow>=10.2.0
